+++ src/Oro/Bundle/NavigationBundle/Configuration/Definition/MenuTreeBuilder.php
@@ -1,15 +1,14 @@
 <?php
 
-namespace Oro\Bundle\NavigationBundle\Config\Definition\Builder;
+namespace Oro\Bundle\NavigationBundle\Configuration\Definition;
 
 use Symfony\Component\Config\Definition\Builder\NodeBuilder;
 
+/**
+ * The builder for menu node.
+ */
 class MenuTreeBuilder extends NodeBuilder
 {
-
-    /**
-     * Constructor
-     */
     public function __construct()
     {
         parent::__construct();
@@ -18,10 +17,11 @@ class MenuTreeBuilder extends NodeBuilder
     }
 
     /**
-     * Creates a child menu node
+     * Creates a child menu node.
      *
-     * @param  string             $name The name of the node
-     * @return MenuNodeDefinition The child node
+     * @param string $name The name of the node
+     *
+     * @return MenuNodeDefinition The menu node
      */
     public function menuNode($name)
     {